more crashes were fixed

+++ main.py
@@ -207,7 +207,7 @@ if __name__ == "__main__":
 
     pfpRatio = profileImages[0].get_size()
     tempR = (sectionHeight * .9)/pfpRatio[1] # using y axis
-    pfpSize = (pfpRatio[0] * tempR, pfpRatio[1] * tempR)
+    pfpSize = (int(pfpRatio[0] * tempR), int(pfpRatio[1] * tempR))
     for i in range(len(profileImages)): # ok wait, these transforms need to be 3:2 ratio
         profileImages[i] = pygame.transform.scale(profileImages[i], pfpSize)
         
@@ -381,11 +381,11 @@ if __name__ == "__main__":
                 (width * (1-(border*2))) / len(requestGroups),
                 height * .5
             ]
-            textBackgrounds[0] = pygame.transform.scale(textBackgrounds[0], (midBoxes[0], midBoxes[1]))
+            textBackgrounds[0] = pygame.transform.scale(textBackgrounds[0], (int(midBoxes[0]), int(midBoxes[1])))
             for i in range(len(requestGroups)):
                 tempCoords = (
                     (width * border * (i)) + (midBoxes[0] * i), height * .15,
-                    midBoxes[0], midBoxes[1]
+                    int(midBoxes[0]), int(midBoxes[1])
                 )
                 window.blit(textBackgrounds[0], (tempCoords[0], tempCoords[1])) # alright, these display correctly... hooray
                 
@@ -395,7 +395,7 @@ if __name__ == "__main__":
 
             sectionSize = (width*(1-(border*2)))/len(FOCUS_GROUPS) # side buffer not included
             sectionHeight = height * .35
-            textBackgrounds[1] = pygame.transform.scale(textBackgrounds[1], (sectionSize, sectionHeight))
+            textBackgrounds[1] = pygame.transform.scale(textBackgrounds[1], (int(sectionSize), int(sectionHeight)))
             
             for i in range(len(FOCUS_GROUPS)): # the background box
                 tempCoords = (
